stockEngine/requirements.txt
asgiref==3.2.3
beautifulsoup4==4.9.3
certifi==2019.11.28
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.8
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-registration==3.1
django-rest-auth==0.9.5
django-webpack-loader==0.6.0
djangorestframework==3.11.0
gunicorn==19.9.0
idna==2.8
lxml==4.6.2
multitasking==0.0.9
numpy==1.20.1
oauthlib==3.1.0
pandas==1.2.2
PyMySQL==1.0.2
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.2
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.8
whitenoise==5.2.0
yfinance==0.1.55
